Move SkinnableTestScene to main project for consumption by other rulesets
