+++ osu.Game.Rulesets.Osu.Tests/TestSceneGameplayCursor.cs
@@ -10,6 +10,7 @@ using osu.Framework.Testing.Input;
 using osu.Game.Configuration;
 using osu.Game.Rulesets.Osu.UI.Cursor;
 using osu.Game.Screens.Play;
+using osu.Game.Tests.Visual;
 using osuTK;
 
 namespace osu.Game.Rulesets.Osu.Tests
