+++ osu.Game.Rulesets.Osu.Tests/TestSceneHitCircle.cs
@@ -14,6 +14,7 @@ using osu.Game.Rulesets.Mods;
 using System.Linq;
 using NUnit.Framework;
 using osu.Game.Rulesets.Scoring;
+using osu.Game.Tests.Visual;
 
 namespace osu.Game.Rulesets.Osu.Tests
 {
